+++ packages/config/src/v1/types.ts
@@ -1,0 +1,26 @@
+/**
+ * Type declarations for @vercel/config
+ *
+ * Usage:
+ * /// <reference types="@vercel/config/v1/types" />
+ */
+
+export type {
+  VercelConfig,
+  Framework,
+  FunctionConfig,
+  CronJob,
+  GitDeploymentConfig,
+  GitConfig,
+  GithubConfig,
+  ImageConfig,
+  Header,
+  Condition,
+  Redirect,
+  Rewrite,
+  HeaderRule,
+  RouteType,
+  WildcardDomain,
+  BuildConfig,
+  BuildItem,
+} from '../types';