+++ packages/fastify/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vercel/fastify
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.5.16
+
+## 0.1.17
+
+### Patch Changes
+
+- Export prepareCache function for backend builders ([#14454](https://github.com/vercel/vercel/pull/14454))
+
 ## 0.1.16
 
 ### Patch Changes