+++ packages/h3/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vercel/h3
 
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.5.16
+
+## 0.1.23
+
+### Patch Changes
+
+- Export prepareCache function for backend builders ([#14454](https://github.com/vercel/vercel/pull/14454))
+
 ## 0.1.22
 
 ### Patch Changes